Add tests for new driver code
